meson: Attempt to fix -Duninstalled_build=true mode
Building in this mode and running ./shell/rhythmbox still fails, with
an error like this:

    /usr/lib64/python3.9/site-packages/gi/module.py:163: Warning: cannot register existing type 'RBShell'
     g_type = info.get_g_type()

    (rhythmbox:56952): GLib-CRITICAL **: 19:19:36.617: g_once_init_leave: assertion 'result != 0' failed

The problem is that libpeas has the wrong library path when loading
plugins. It tries to load symbols from /usr/lib64/librhythmboxcore.so
and these conflict with those from ./shell/librhythmboxcore.so which
is already loaded.

We already set rpath to $ORIGIN in the `rhythmbox` binary, so I'm not
sure what more can be done. Setting LD_LIBRARY_PATH manually also
doesn't get it to work.